+++ src/test/java/io/appform/dropwizard/discovery/bundle/ServiceDiscoveryBundleCustomHostPortTest.java
@@ -99,7 +99,7 @@ class ServiceDiscoveryBundleCustomHostPortTest {
     void setup() throws Exception {
         val connector = mock(LocalConnector.class);
         when(connector.getConnectionFactory(Mockito.anyString())).thenReturn(null);
-        Connector[] connectors = new Connector[1];
+        val connectors = new Connector[1];
         connectors[0] = connector;
         when(server.getConnectors()).thenReturn(connectors);