+++ src/test/java/io/appform/dropwizard/discovery/bundle/ServiceDiscoveryBundleDwMonitorTest.java
@@ -125,7 +125,7 @@ class ServiceDiscoveryBundleDwMonitorTest {
 
         val connector = mock(LocalConnector.class);
         when(connector.getConnectionFactory(Mockito.anyString())).thenReturn(null);
-        Connector[] connectors = new Connector[1];
+        val connectors = new Connector[1];
         connectors[0] = connector;
         when(server.getConnectors()).thenReturn(connectors);