+++ src/test/java/io/appform/dropwizard/discovery/bundle/ServiceDiscoveryBundleHierarchicalSelectorTest.java
@@ -97,7 +97,7 @@ class ServiceDiscoveryBundleHierarchicalSelectorTest {
     void setup() throws Exception {
         val connector = mock(LocalConnector.class);
         when(connector.getConnectionFactory(Mockito.anyString())).thenReturn(null);
-        Connector[] connectors = new Connector[1];
+        val connectors = new Connector[1];
         connectors[0] = connector;
         when(server.getConnectors()).thenReturn(connectors);