+++ src/test/java/io/appform/dropwizard/discovery/bundle/ServiceDiscoveryBundleTest.java
@@ -96,7 +96,7 @@ class ServiceDiscoveryBundleTest {
     void setup() throws Exception {
         val connector = mock(LocalConnector.class);
         when(connector.getConnectionFactory(Mockito.anyString())).thenReturn(null);
-        Connector[] connectors = new Connector[1];
+        val connectors = new Connector[1];
         connectors[0] = connector;
         when(server.getConnectors()).thenReturn(connectors);