+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [2.0.28-6]
-- Added a criteriaResolver and introduced a TransportType resolver as subType
+- Added a criteriaResolver and introduced a PortScheme resolver as subType
 - Upgraded ranger version to 1.0-RC10
 
 ## [2.0.28-5]